Move types used in ipc to ipc-types.ts
