+++ gui/src/shared/ipc-types.ts
@@ -1,0 +1,10 @@
+export interface ICurrentAppVersionInfo {
+  gui: string;
+  daemon: string;
+  isConsistent: boolean;
+  isBeta: boolean;
+}
+
+export interface IWindowShapeParameters {
+  arrowPosition?: number;
+}
